portfolio/requirements.txt
streamlit==1.50.0
matplotlib==3.9.4
plotly==6.3.1
pandas              
numpy                    
watchdog            
streamlit-mermaid
pyyaml